+++ pages/login.tsx
@@ -1,6 +1,11 @@
 import React, { useState } from 'react';
 import { FieldValues, useForm } from 'react-hook-form';
 import { getSession, GetSessionParams, signIn } from 'next-auth/react';
+import { Media } from '@/lib/media';
+import Link from 'next/link';
+import Image from 'next/image';
+import LeftDisplay from '@/components/Register/LeftDisplay';
+import MobileLogin from '@/components/mobile/MobileLogin/MobileLogin';
 import {
   CreateButton,
   Container,
@@ -12,91 +17,21 @@ import {
   LoginHeader,
   RightContainer,
   SubmitButton,
+  PasswordWrapper,
+  Eye,
+  ForgotPassword,
+  MobileContainer,
+  MobileImageContainer,
+  MobileTextContainer,
+  MobileText,
+  MobileLoginButton,
 } from '@/styles/login.styles';
-import LeftDisplay from '@/components/Register/LeftDisplay';
-import MobileLogin from '@/components/mobile/MobileLogin/MobileLogin';
-import Link from 'next/link';
-import { Media } from '@/lib/media';
-import styled from 'styled-components';
-import Image from 'next/image';
-
-const PasswordWrapper = styled.div`
-  position: relative;
-  display: flex;
-`;
-
-const Eye = styled.i`
-  position: absolute;
-  top: 13%;
-  right: 4%;
-`;
-
-const ForgotPassword = styled.div`
-  margin: -10px 10px 0 auto;
-  font-size: 15px;
-  line-height: 18px;
-  color: #6d6d6d;
-`;
-
-interface Props {
-  hover?: boolean;
-}
-
-const MobileContainer = styled.div`
-  height: 100vh;
-  padding: 8.6% 14% 6% 14%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 25px;
-  line-height: 30.26px;
-`;
-
-const MobileImageContainer = styled.div`
-  position: relative;
-  width: 73vw;
-  height: 46vh;
-`;
-
-const MobileTextContainer = styled.div`
-  margin: auto;
-`;
-
-const MobileText = styled.p<Props>`
-  text-align: center;
-  &:last-child {
-    font-size: 16px;
-    line-height: 19.36px;
-  }
-  &:hover {
-    cursor: ${props => (props.hover ? `pointer` : `auto`)};
-  }
-`;
-
-const MobileLoginButton = styled.button`
-  background: #6d6d6d;
-  border: none;
-  border-radius: 10px;
-  font-size: 1rem;
-  line-height: 1rem;
-  color: white;
-  width: 296px;
-  height: 53px;
-  &:hover {
-    cursor: pointer;
-  }
-`;
 
 const LoginPage = () => {
   // state for going to mobile login page
   const [onMobileLogin, setOnMobileLogin] = useState(false);
 
-  // state for showing psasword
+  // state for showing password
   const [passwordShown, setPasswordShown] = useState(false);
 
   // React hook form.
@@ -121,41 +56,15 @@ const LoginPage = () => {
 
   return (
     <>
-      <Media lessThan="sm">
-        {onMobileLogin ? (
-          <MobileLogin />
-        ) : (
-          <MobileContainer>
-            <MobileImageContainer>
-              <Image src={'/bookemkids.png'} alt="BookEm Background" fill />
-            </MobileImageContainer>
-
-            <MobileTextContainer>
-              <MobileText>
-                Welcome to the Book&apos;em Volunteer Portal
-              </MobileText>
-
-              <MobileText>
-                Share the joy of reading and book ownership.
-              </MobileText>
-            </MobileTextContainer>
-
-            <MobileLoginButton onClick={() => setOnMobileLogin(true)}>
-              Log in
-            </MobileLoginButton>
-
-            <Link href={'/register'}>
-              <MobileText hover>Or sign up</MobileText>
-            </Link>
-          </MobileContainer>
-        )}
-      </Media>
+      {/* Desktop */}
       <Media greaterThanOrEqual="sm">
         <Container>
           <LeftDisplay />
+
           <RightContainer>
             <ContentContainer>
               <LoginHeader>Logging you in</LoginHeader>
+
               <LoginForm
                 id="loginForm"
                 onSubmit={handleSubmit(data => handleLogin(data))}>
@@ -163,6 +72,7 @@ const LoginPage = () => {
                   {...register('email', { required: true })}
                   placeholder="Email or username"
                 />
+
                 <PasswordWrapper>
                   <Input
                     {...register('password', { required: true })}
@@ -194,17 +104,50 @@ const LoginPage = () => {
                 {errors.email && <span>Email is required</span>}
                 {errors.password && <span>Password is required</span>}
               </LoginForm>
+
               <SubmitButton form="loginForm" type="submit" value="Log in" />
             </ContentContainer>
 
             <Footer>
               <LittleText>New here? Come join us!</LittleText>
+
               <Link href={'/register'}>
                 <CreateButton>Sign up</CreateButton>
               </Link>
             </Footer>
           </RightContainer>
         </Container>
+      </Media>
+
+      {/* Mobile */}
+      <Media lessThan="sm">
+        {onMobileLogin ? (
+          <MobileLogin />
+        ) : (
+          <MobileContainer>
+            <MobileImageContainer>
+              <Image src={'/bookemkids.png'} alt="BookEm Background" fill />
+            </MobileImageContainer>
+
+            <MobileTextContainer>
+              <MobileText>
+                Welcome to the Book&apos;em Volunteer Portal
+              </MobileText>
+
+              <MobileText>
+                Share the joy of reading and book ownership.
+              </MobileText>
+            </MobileTextContainer>
+
+            <MobileLoginButton onClick={() => setOnMobileLogin(true)}>
+              Log in
+            </MobileLoginButton>
+
+            <Link href={'/register'}>
+              <MobileText hover>Or sign up</MobileText>
+            </Link>
+          </MobileContainer>
+        )}
       </Media>
     </>
   );
